fix warning on amd64 compile (cherry picked from commit 7ad68c6621b272c51af12ca637f1036b2cd07318)

+++ drbd/drbd_receiver.c
@@ -2916,7 +2916,7 @@ STATIC void drbd_disconnect(struct drbd_conf *mdev)
 		for (h = mdev->ee_hash; h < mdev->ee_hash + mdev->ee_hash_s; h++)
 			if (h->first)
 				ERR("ASSERT FAILED ee_hash[%u].first == %p, expected NULL\n",
-						h - mdev->ee_hash, h->first);
+						(int)(h - mdev->ee_hash), h->first);
 		kfree(mdev->ee_hash);
 		mdev->ee_hash = NULL;
 		mdev->ee_hash_s = 0;
@@ -2925,7 +2925,7 @@ STATIC void drbd_disconnect(struct drbd_conf *mdev)
 		for (h = mdev->tl_hash; h < mdev->tl_hash + mdev->tl_hash_s; h++)
 			if (h->first)
 				ERR("ASSERT FAILED tl_hash[%u] == %p, expected NULL\n",
-						h - mdev->tl_hash, h->first);
+						(int)(h - mdev->tl_hash), h->first);
 		kfree(mdev->tl_hash);
 		mdev->tl_hash = NULL;
 		mdev->tl_hash_s = 0;
